Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -6,7 +6,7 @@ int main(){
     scanf("%d" , &b);
     scanf("%d" , &c);
     int sum = a + b + c;
-    int average = sum / 3;
-    printf("Average: %d\n" , average);
+    float average = sum / 3;
+    printf("Average: %f\n" , average);
     return 0;
 }
